+++ tests/test_app.py
@@ -8,7 +8,7 @@ import os
 import uuid
 
 import pytest
-from pvsite_datamodel.sqlmodels import ForecastSQL, ForecastValueSQL, MLModelSQL, SiteAssetType
+from pvsite_datamodel.sqlmodels import ForecastSQL, ForecastValueSQL, MLModelSQL
 
 from india_forecast_app.app import (
     app,
@@ -39,26 +39,6 @@ def test_get_sites(db_session, sites):
         assert isinstance(site.site_uuid, uuid.UUID)
         assert sites[0].asset_type.name == "pv"
         assert sites[1].asset_type.name == "wind"
-
-
-def test_get_generation_data(db_session, sites, generation_db_values, init_timestamp):
-    """Test for correct generation data"""
-
-    # Test only checks for wind data as solar data not ready yet
-    gen_sites = [s for s in sites if s.asset_type == SiteAssetType.wind][0:1]  # 1 site
-    gen_data = get_generation_data(db_session, gen_sites, timestamp=init_timestamp)
-    gen_df, gen_meta = gen_data["data"], gen_data["metadata"]
-
-    # Check for 5 (non-null) generation values
-    assert len(gen_df) == 5
-    assert not gen_df["0"].isnull().any()  # 0 is the ml_id/system_id of the wind site
-
-    # Check first and last timestamps are correct
-    assert gen_df.index[0] == init_timestamp - dt.timedelta(hours=1)
-    assert gen_df.index[-1] == init_timestamp
-
-    # Check for expected metadata
-    assert len(gen_meta) == 1
 
 
 @pytest.mark.parametrize("asset_type", ["pv", "wind"])
